+++ src/components/modules/popup/details/ImagesUnit.tsx
@@ -1,6 +1,5 @@
 import type {ImageData} from '_scripts/imagesExtractor'
 import {ROUTES} from '@/lib/routes'
-import {cn} from '@/lib/utils'
 
 import Unit from '~/UI/Unit'
 import ImageCell from '~/UI/ImageCell'
@@ -25,9 +24,9 @@ export default function ImagesUnit({data}: {data: ImageData[] | undefined}) {
   }
 
   return (
-    <Unit token="images" className={cn('flex gap-[7px]')}>
+    <Unit token="images" className="flex gap-2">
       {imagesData.slice(0, 4).map(({src, type}) => (
-        <ImageCell source={src} type={type} key={src} />
+        <ImageCell view="details" source={src} type={type} key={src} />
       ))}
 
       <ExpandButton to={ROUTES.images} />