+++ src/components/modules/popup/images/Gallery.tsx
@@ -1,11 +1,26 @@
-import type {ImagesResponse} from '_modules/popup/details/Images'
+import type {ImagesResponse, ImageMode} from '_modules/popup/details/Images'
+
+import {cn} from '@/lib/utils'
 
 import ImageCell from '~/UI/ImageCell'
 
-export default function Gallery({data}: {data: ImagesResponse | undefined}) {
+export default function Gallery({data, format, className}: {data: ImagesResponse | undefined; format: ImageMode; className?: string}) {
+  const filteredImages = data?.images?.filter(({type}) => {
+    switch (format) {
+      case 'all':
+        return true
+      case 'img':
+        return type === 'img' || type === 'bg-image'
+      case 'icon':
+        return type === 'icon'
+      default:
+        return false
+    }
+  })
+
   return (
-    <section data-section="gallery-images" className="grid grid-cols-2 gap-2">
-      {data?.images?.map(({src, type}, idx) => <ImageCell source={src} className="h-36 bg-control/40" type={type} key={idx} />)}
+    <section data-section="gallery-images" className={cn('grid grid-cols-3 gap-2', className)}>
+      {filteredImages?.map(({src, type}, idx) => <ImageCell view="gallery" source={src} type={type} key={idx} />)}
     </section>
   )
 }