+++ src/modules/popup/details/Colors.tsx
@@ -2,8 +2,11 @@ import type {ColorData} from '_scripts/colorsExtractor'
 import {useEffect, useState} from 'react'
 
 import Layout from '~/Global/Layout'
+import Controls, {ControlsTabItem, CONTROLS_GAP} from '~/UI/Controls'
 import Palette from '~~popup/colors/Palette'
-import Controls, {type ColorMode} from '~~popup/colors/Controls'
+
+const modes = ['hex', 'rgb', 'hsl'] as const
+export type ColorMode = (typeof modes)[number]
 
 export type ColorsResponse = {
   colors: ColorData[]
@@ -28,9 +31,14 @@ export function Colors() {
   }
 
   return (
-    <Layout className="space-y-3.5">
-      <Controls selectedMode={colorFormat} onModeChange={handleFormatChange} />
-      <Palette data={colorsData} format={colorFormat} />
+    <Layout className="space-y-3.5 relative">
+      <Controls>
+        {modes.map((mode) => (
+          <ControlsTabItem item={mode} isSelected={mode === colorFormat} onClick={() => handleFormatChange(mode)} key={mode} />
+        ))}
+      </Controls>
+
+      <Palette data={colorsData} format={colorFormat} className={CONTROLS_GAP} />
     </Layout>
   )
 }