+++ src/modules/popup/details/Details.tsx
@@ -25,9 +25,9 @@ export function Details() {
       <ColorsUnit data={units?.colors} />
       <ImagesUnit data={units?.images} />
 
-      <p className="text-xs text-gray/50 text-center leading-none group">
+      <p className="text-xs text-gray-dark text-center leading-none group">
         we use data for advertising. check{' '}
-        <a className="text-gray/60 underline underline-offset-1 group-hover:no-underline" href="https://github.com/bozzhik/snable/blob/main/PRIVACY_POLICY.md" target="_blank">
+        <a className="text-gray-medium underline underline-offset-1 group-hover:no-underline" href="https://github.com/bozzhik/snable/blob/main/PRIVACY_POLICY.md" target="_blank">
           privacy policy
         </a>
       </p>